+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/util/PaginationFilter.java
@@ -1,16 +1,14 @@
 package uk.ac.bbsrc.tgac.miso.core.util;
 
-import static uk.ac.bbsrc.tgac.miso.core.util.LimsUtils.isStringEmptyOrNull;
-
-import java.util.*;
+import java.util.Collection;
+import java.util.Date;
+import java.util.EnumSet;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import org.joda.time.DateTime;
-
-import sun.awt.image.ImageWatched;
 import uk.ac.bbsrc.tgac.miso.core.data.Identifiable;
 import uk.ac.bbsrc.tgac.miso.core.data.Pool;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.PoolImpl;
@@ -21,10 +19,6 @@ import uk.ac.bbsrc.tgac.miso.core.data.type.PlatformType;
 
 @SuppressWarnings("squid:S1604")
 public abstract interface PaginationFilter {
-  public final static List<AgoMatcher> AGO_MATCHERS = Arrays.asList(new AgoMatcher("h(|ours?)", 3600),
-      new AgoMatcher("d(|ays?)", 3600 * 24));
-
-  public static final Pattern WHITESPACE = Pattern.compile("\\s+(?=([^\"]*\"[^\"]*\")*[^\"]*$)");
 
   public static PaginationFilter archived(boolean isArchived) {
     return new PaginationFilter() {
@@ -104,10 +98,7 @@ public abstract interface PaginationFilter {
     };
   }
 
-  public static PaginationFilter distributed(String maybeDistributionDates) {
-    if (!isStringEmptyOrNull(maybeDistributionDates)) {
-      return parseDate(maybeDistributionDates, DateType.DISTRIBUTED);
-    }
+  public static PaginationFilter distributed() {
     return new PaginationFilter() {
 
       @Override
@@ -259,167 +250,6 @@ public abstract interface PaginationFilter {
         sink.restrictPaginationBySequencingParametersName(item, name, errorHandler);
       }
     };
-  }
-
-  /**
-   * Search terms are documented in miso-web/src/main/webapp/scripts/list.js
-   */
-  public static PaginationFilter[] parse(String request, String currentUser, Consumer<String> errorHandler) {
-    return WHITESPACE.splitAsStream(request).map(x -> {
-      x = x.replace("\"", "");
-      if (x.isEmpty()) return null;
-      if (x.contains(":")) {
-        String[] parts = x.split(":", 2);
-        switch (parts[0]) {
-        case "is":
-        case "has":
-          switch (parts[1].toLowerCase()) {
-          case "fulfilled":
-            return fulfilled(true);
-          case "active":
-          case "order":
-          case "ordered":
-          case "unfulfilled":
-            return fulfilled(false);
-          case "unknown":
-            return health(HealthType.Unknown);
-          case "complete":
-          case "completed":
-            return health(HealthType.Completed);
-          case "failed":
-            return health(HealthType.Failed);
-          case "started":
-            return health(HealthType.Started);
-          case "stopped":
-            return health(HealthType.Stopped);
-          case "running":
-            return health(HealthType.Running);
-          case "incomplete":
-            return health(EnumSet.of(HealthType.Running, HealthType.Started, HealthType.Stopped));
-          case "ghost":
-            return ghost(true);
-          case "real":
-            return ghost(false);
-          default:
-            errorHandler.accept("No filter for " + x);
-            return null;
-          }
-        case "created":
-        case "createdon":
-          return parseDate(parts[1], DateType.CREATE);
-        case "changed":
-        case "modified":
-        case "updated":
-        case "changedon":
-        case "modifiedon":
-        case "updatedon":
-          return parseDate(parts[1], DateType.UPDATE);
-        case "received":
-        case "recieved":
-        case "receivedon":
-        case "recievedon":
-          return parseDate(parts[1], DateType.RECEIVE);
-        case "createdby":
-        case "creator":
-        case "creater":
-          return parseUser(parts[1], currentUser, true);
-        case "changedby":
-        case "modifier":
-        case "updater":
-          return parseUser(parts[1], currentUser, false);
-        case "platform":
-          try {
-            return platformType(PlatformType.valueOf(parts[1].toUpperCase()));
-          } catch (IllegalArgumentException e) {
-            errorHandler.accept("Invalid platform: " + parts[1]);
-            return null;
-          }
-        case "index":
-          return index(parts[1]);
-        case "class":
-          return sampleClass(parts[1]);
-        case "external":
-        case "ext":
-        case "extern":
-          return external(parts[1]);
-        case "institute":
-        case "inst":
-          return institute(parts[1]);
-        case "box":
-          return box(parts[1]);
-        case "kitname":
-          return kitName(parts[1]);
-        case "subproject":
-          return subproject(parts[1]);
-        case "sequencingparameters":
-        case "parameters":
-        case "params":
-          return sequencingParameters(parts[1]);
-        case "groupid":
-          return groupId(parts[1]);
-        case "distributed":
-          return distributed(parts[1]);
-        case "distributedto":
-          return distributedTo(parts[1]);
-        }
-      }
-      return query(x);
-    }).filter(Objects::nonNull).toArray(PaginationFilter[]::new);
-  }
-
-  static PaginationFilter parseDate(String text, DateType type) {
-    DateTime start;
-    DateTime end;
-    switch (text.toLowerCase()) {
-    case "now":
-    case "hour":
-    case "thishour":
-    case "lasthour":
-      end = new DateTime();
-      start = end.minusHours(1);
-      break;
-    case "today":
-      start = new DateTime().withTimeAtStartOfDay();
-      end = start.plusDays(1);
-      break;
-    case "yesterday":
-      end = new DateTime().withTimeAtStartOfDay();
-      start = end.minusDays(1);
-      break;
-    case "thisweek":
-      end = new DateTime().withTimeAtStartOfDay().plusDays(1);
-      start = end.withDayOfWeek(1);
-      break;
-    case "lastweek":
-      end = new DateTime().withTimeAtStartOfDay().plusDays(1).minusWeeks(1);
-      start = end.withDayOfWeek(1);
-      break;
-    default:
-      int ago = 0;
-      for (int i = 0; i < AGO_MATCHERS.size() && ago == 0; i++) {
-        ago = AGO_MATCHERS.get(i).secondsAgoIfmatches(text);
-      }
-      if (ago > 0) {
-        end = new DateTime();
-        start = end.minusSeconds(ago);
-      } else {
-        try {
-          DateTime d = DateTime.parse(text);
-          start = d.withTimeAtStartOfDay();
-          end = d.withTimeAtStartOfDay().plusDays(1);
-        } catch (IllegalArgumentException e) {
-          return null;
-        }
-      }
-    }
-    return date(start.toDate(), end.toDate(), type);
-  }
-
-  static PaginationFilter parseUser(String username, String currentUser, boolean creator) {
-    if (username.equalsIgnoreCase("me")) {
-      return user(currentUser, creator);
-    }
-    return user(username, creator);
   }
 
   public static PaginationFilter pending() {